+++ tests/Neba.TestFactory/Seasons/HighBlockAwardFactory.cs
@@ -3,7 +3,7 @@ using Bogus;
 using Neba.Domain.Bowlers;
 using Neba.Domain.Seasons;
 
-namespace Neba.TestFactory.Awards;
+namespace Neba.TestFactory.Seasons;
 
 public static class HighBlockAwardFactory
 {
